src/requirements.txt
waitress
paramiko
python-socketio
sqlalchemy
flask
pexpect
pandas
pymu
PyMySQL
flask-cors
sqlmap
rut-chile
rake-nltk
flask-restx
bs4
requests
cloudscraper
fastparquet
asyncpg
pangres
flatten_json
webdriver-manager
llama-cpp-python
psycopg2-binary